Organize project with mvn module
