+++ productview-persistence/src/main/java/com/catbagan/productview/persistence/repositories/ProductRepository.java
@@ -1,0 +1,14 @@
+package com.catbagan.productview.persistence.repositories;
+
+import org.springframework.data.repository.CrudRepository;
+
+import com.catbagan.productview.persistence.models.Product;
+
+/**
+ * 
+ * @author jenseric
+ *
+ */
+public interface ProductRepository extends CrudRepository<Product, Long> {
+
+}
