+++ productview-web/pom.xml
@@ -1,0 +1,28 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+	<parent>
+		<groupId>com.catbagan</groupId>
+		<artifactId>productview</artifactId>
+		<version>0.0.1-SNAPSHOT</version>
+	</parent>
+	<artifactId>productview-web</artifactId>
+	<name>productview-web</name>
+	<description>web module</description>
+	
+	<dependencies>
+		<dependency>
+			<groupId>com.catbagan</groupId>
+			<artifactId>productview-service</artifactId>
+			<version>0.0.1-SNAPSHOT</version>
+		</dependency>
+		
+				
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-web</artifactId>
+			<version>${spring.boot.version}</version>
+		</dependency>
+		
+	</dependencies>
+</project>
