+++ productview-web/src/main/java/com/catbagan/productview/ProductviewApplication.java
@@ -1,0 +1,17 @@
+package com.catbagan.productview;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+/**
+ * main 
+ * @author Jenseric.Catbagan
+ *
+ */
+public class ProductviewApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ProductviewApplication.class, args);
+	}
+}
