+++ productview-web/src/test/java/com/catbagan/productview/web/ProductviewApplicationTests.java
@@ -1,0 +1,16 @@
+package com.catbagan.productview.web;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class ProductviewApplicationTests {
+
+	@Test
+	public void contextLoads() {
+	}
+
+}
