--- src/main/java/com/catbagan/productview/ProductviewApplication.java
@@ -1,12 +1,0 @@
-package com.catbagan.productview;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ProductviewApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ProductviewApplication.class, args);
-	}
-}
