--- src/main/java/com/catbagan/productview/controllers/ProductController.java
@@ -1,53 +1,0 @@
-package com.catbagan.productview.controllers;
-
-import java.util.Collection;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RestController;
-
-import com.catbagan.productview.models.Product;
-import com.catbagan.productview.services.ProductService;
-
-/**
- * 
- * @author jenseric
- *
- */
-@RestController
-public class ProductController {
-
-	@Autowired
-	private ProductService productService;
-	
-	@GetMapping("/products")
-	public Collection<Product> getAllProduct() {
-		return getProductService().getAllProduct();
-	}
-	
-	@GetMapping("/products/{id}")
-	public Product getProduct(final @PathVariable long id) {
-		final Product product = getProductService().findById(id);
-		return product;
-	}
-	
-	
-
-	/**
-	 * @return the productService
-	 */
-	public ProductService getProductService() {
-		return productService;
-	}
-
-	/**
-	 * @param productService the productService to set
-	 */
-	public void setProductService(ProductService productService) {
-		this.productService = productService;
-	}
-
-	
-
-}
