--- src/main/java/com/catbagan/productview/repositories/ProductRepository.java
@@ -1,14 +1,0 @@
-package com.catbagan.productview.repositories;
-
-import org.springframework.data.repository.CrudRepository;
-
-import com.catbagan.productview.models.Product;
-
-/**
- * 
- * @author jenseric
- *
- */
-public interface ProductRepository extends CrudRepository<Product, Long> {
-
-}
